+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Added
+
+- Yaw rate safety check has been added with a maximum yaw rate threshold.
+
+### Fixed
+
+- Fixed when acceleration warnings appear as a safety check.
+
 ## [4.1.0] - 2026-01-14
 
 ### Added
 
-- Optional acceleration plot is added to the validation PDF file (PRO only)
+- Optional acceleration plot is added to the validation PDF file (PRO only).
 
 ### Changed
 
@@ -18,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Yaw and pyro information can now be part of the validation PDF file (PRO only)
+- Yaw and pyro information can now be part of the validation PDF file (PRO only).
 
 - A warning is now printed when importing a formation from an external file that
   contains no trajectories or light programs, instead of failing with an error.